Add LocalDateUtils class
Purpose: Abstract all functions that interacts with LocalDate into a
single class
