+++ src/test/java/seedu/address/model/util/LocalDateUtilsTest.java
@@ -1,0 +1,4 @@
+package seedu.address.model.util;
+
+public class LocalDateUtilsTest {
+}
